sum them all arguments

+++ 0x10-variadic_functions/0-sum_them_all.c
@@ -11,8 +11,8 @@
 int sum_them_all(const unsigned int n, ...)
 {
 	va_list args;
-	size_t sum = 0;
-	double i;
+	int sum = 0;
+	size_t i;
 
 	if (n == 0)
 		return (0);
